+++ src/test/java/model/card/DeckTest.java
@@ -1,14 +1,8 @@
-package model;
+package model.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import model.card.Card;
-import model.card.CardRank;
-import model.card.CardSuit;
-import model.card.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
