+++ src/test/java/model/participant/DealerTest.java
@@ -1,11 +1,10 @@
-package model;
+package model.participant;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import model.card.Card;
 import model.card.CardRank;
 import model.card.CardSuit;
-import model.participant.Dealer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
