+++ src/test/java/model/participant/PlayerTest.java
@@ -1,11 +1,10 @@
-package model;
+package model.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import model.card.Card;
 import model.card.CardRank;
 import model.card.CardSuit;
-import model.participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
