+++ src/test/java/model/participant/PlayersTest.java
@@ -1,9 +1,8 @@
-package model;
+package model.participant;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import model.participant.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
